The framework has been overhauled to be multithreaded. This leads to up to a 40% performance improvement with certain filters, and should prevent blocking of the interface when processing video. However, it may break things.

+++ framework/Source/GPUImageKuwaharaFilter.m
@@ -104,7 +104,7 @@ NSString *const kGPUImageKuwaharaFragmentShaderString = SHADER_STRING
 {
     _radius = newValue;
     
-    [self setFloat:_radius forUniform:radiusUniform program:filterProgram];
+    [self setInteger:_radius forUniform:radiusUniform program:filterProgram];
 }
 
 @end
